requirements.txt
appium-python-client
selenium
pytest
pytest-html
allure-pytest
openpyxl
